requirements.txt
promptflow
promptflow-tools
python-dotenv>=0.10.3
jinja2
mlflow>=2.7.1
azure-ai-ml>=1.11.0
azure-identity>=1.15.0
azureml-mlflow>=1.51
keyrings.alt
pyyaml
